test(interp): make the ARM64 JIT tests an importing client
interp/jit_arm64_test.go declared package interp and was the largest remaining
violation of docs/coding-patterns.md §12.2. It also would not survive the JIT
moving to its own package, since every reference to interp's private state
would stop compiling.

Almost all of it was already public-contract clean. The only private access
across twenty of its twenty-four tests was reading the refcount array directly,
and interp_test.go already had the public replacement: refCounts, built on the
exported RefCount and HeapLen. Those tests now run as package interp_test.

The four that cannot be expressed publicly move to interp/jit_test.go, still
package interp:

  TestCompiler_Compile calls the compiled callable directly with journalPtr,
  bypassing dispatch to assert the native trap encoding. That encoding is the
  native-to-interpreter contract itself and has no public projection.

  TestARM64_Backedge reads install bookkeeping - whether a root was tried,
  whether an entry is installed. No metric separates "attempted, not installed"
  from "not attempted".

  TestARM64_StackReserve holds three hand-synced constants together against a
  literal in abi_arm64.s. Two of the three are unexported sizing constants.

  TestARM64_Flush hand-builds lowering IR to prove a dirty carried local emits
  no materialization.

Each carries the reason it is parked there and the condition that removes it,
and docs/testing.md now records the three remaining white-box files rather than
claiming every test package is an importing client.

One coupling is now mirrored rather than shared: the loop safepoint budget the
yield subtest sizes its loop against. The copy is commented with the drift it
introduces.

No test or subtest was lost - 24 top-level tests and 104 subtests before and
after.
